feat: use SVGR to import SVG as React component
This commit introduces vite-plugin-svgr to handle SVG imports as React components.
This allows us to use SVGs directly as components, improving code readability
and maintainability. The placeholder app icon is now imported as a React
component using SVGR. Additionally, the icon property in the app content
files is updated to be an empty string.